requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
supabase==2.4.0
python==3.13
